Move image handler/router to its own file
